:recycle: Address review points on recurring transaction changes
